Fix NML floor division, regenerate all data, correct paper tables
Critical fix: template_bits_nml used float division (steps/period) instead
of floor division (steps//period) as stated in the paper. This changes
period selection in marginal cases.

Data corrections from regenerated outputs:
- S11/B37 at T=60 selects period 2 (not 4) with 7-bit margin
- S25/B12 and S11/B37 lose perfect seed consistency (8/10, 7/10)
- S24/B11 400-step verification now shows period 4 (not 6)
- S37/B11 scaling: 32x32 and 64x64 defect counts changed

Also: mdl_bits type annotation fixed (int -> float), 986-candidate
phantom reference removed from Appendix A, S11/B37 table now shows
all 5 grid sizes, period convergence script added with backing CSV.

+++ src/relative_symmetry_repair/repair_nd.py
@@ -25,7 +25,7 @@ class RelativePeriodicFitND:
     run_length_bits: int
     lz4_bits: int
     template_bits: int = 0
-    mdl_bits: int = 0           # template_bits + run_length_bits
+    mdl_bits: float = 0.0       # template_bits_nml + run_length_bits
     rule_error: float | None = None
 
     def to_record(self) -> dict[str, float | int | None]:
